feature/fluxoMenu - Commit "TODO" finalizados. Refatoração do método Main
