+++ src/DadosTurmas.java
@@ -25,6 +25,27 @@ public class DadosTurmas {
         }
     }
 
+    public static void listarTurmasMatriculado(Aluno alunoEscolhido) {
+        for (Turma turma : listaTurmas) {
+            for (Aluno aluno : turma.listaAlunos) {
+                if (aluno.getIdAluno() == alunoEscolhido.getIdAluno()) {
+//                    System.out.println(turma.getCurso());
+                    System.out.println("ID Turma: [" + turma.getIdTurma() + "]");
+                    System.out.println("Curso: " + turma.getCurso());
+                    System.out.println("Professor: " + turma.getCurso().getProfessor());
+                    System.out.println("Ano de criação: " + turma.getAnoCriacao());
+
+//                    System.out.println("\nAlunos: ");
+//                    for (aluno : turma.getCurso().getAlunos()) {
+//                        System.out.println(" [" + aluno.getIdAluno() + "] - " + aluno.getNome());
+//                    }
+                    System.out.println("----------------------");
+                    System.out.println();
+                }
+            }
+        }
+    }
+
     public static Turma buscarIdTurmas(Scanner scanner) {
         System.out.println("Turmas Cadastradas: ");
         listarTurmas();
@@ -43,6 +64,23 @@ public class DadosTurmas {
         return turmaEscolhida;
     }
 
+    public static Turma buscarIdTurmasMatriculado(Scanner scanner, Aluno alunoEscolhido) {
+        System.out.println("Turmas Cadastradas: ");
+        listarTurmasMatriculado(alunoEscolhido);
+        System.out.println("Digite o id da turma que você quer selecionar");
+
+        int idTurmaEscolhida = scanner.nextInt();
+        scanner.nextLine();
+        Turma turmaEscolhida = null;
+
+        for (Turma turma : listaTurmas) {
+            if (turma != null && turma.getIdTurma() == idTurmaEscolhida) {
+                turmaEscolhida = turma;
+            }
+        }
+
+        return turmaEscolhida;
+    }
 
     public static void removerTurma(Scanner scanner) {
         Turma turma = buscarIdTurmas(scanner);
